Add files via upload
I was challenged to write Pong! clone in a single afternoon. The next day I added a single-player mode with an adjustable AI opponent. At start-up, the user is offered a "Quick Start", or can customise the points target, ball speed, frame rate, and the accuracy of the AI opponent.

Not the most elegant code, but it works!

Pro-tip: leave the sound effects switched off!